feat: Phase 2+ finishing — gRPC streams, IOC persistence, IPv6, clippy
* arqenor-grpc: wire WatchProcesses/WatchFilesystem streams to platform
  providers (new_process_monitor, new_fs_scanner). Add conversion
  helpers core -> proto and 4 tests including a tempfile smoke test.

* arqenor-core/ioc + arqenor-store: persist IOC feeds in SQLite with
  ETag/Last-Modified conditional GET for delta refresh. New
  IocPersistence trait in core, IocSqliteStore impl in store.
  Atomic replacement via single transaction. 49 tests, reboot roundtrip.

* arqenor-platform: add IPv6 support for TCP6 (MIB_TCP6TABLE_OWNER_PID)
  and UDP6 via GetExtendedTcpTable/UdpTable(AF_INET6). Canonical
  [addr]:port formatting. 3 unit tests including simulated row parse.

* arqenor-platform: drop unnecessary usize casts in ntdll_check.rs
  (clippy::unnecessary_cast x 4).

Workspace clippy -D warnings: clean.
Workspace tests: all passing.
Go build/vet/test: clean (proto bindings regenerated locally).

+++ rust/arqenor-platform/src/windows/ntdll_check.rs
@@ -261,7 +261,7 @@ fn parse_exports(pe: &[u8]) -> Option<Vec<ExportEntry>> {
         return None;
     }
 
-    let export_file_off = rva_to_file_offset_raw(pe, pe_off, export_rva as u32)? as usize;
+    let export_file_off = rva_to_file_offset_raw(pe, pe_off, export_rva as u32)?;
 
     if export_file_off + 40 > pe.len() {
         return None;
@@ -272,9 +272,9 @@ fn parse_exports(pe: &[u8]) -> Option<Vec<ExportEntry>> {
     let addr_of_names_rva = read_u32(pe, export_file_off + 32);
     let addr_of_ordinals_rva = read_u32(pe, export_file_off + 36);
 
-    let fn_table = rva_to_file_offset_raw(pe, pe_off, addr_of_functions_rva)? as usize;
-    let name_table = rva_to_file_offset_raw(pe, pe_off, addr_of_names_rva)? as usize;
-    let ord_table = rva_to_file_offset_raw(pe, pe_off, addr_of_ordinals_rva)? as usize;
+    let fn_table = rva_to_file_offset_raw(pe, pe_off, addr_of_functions_rva)?;
+    let name_table = rva_to_file_offset_raw(pe, pe_off, addr_of_names_rva)?;
+    let ord_table = rva_to_file_offset_raw(pe, pe_off, addr_of_ordinals_rva)?;
 
     let mut exports = Vec::with_capacity(num_names);
